Reduce compiler warnings (#29)
This PR contains changes aimed at reducing the amount of compiler warnings generated at build (actually, all the warnings I saw in linux, mac, VS2017 and appveyor).
Changes have been tested on different system configurations, including Mac, Linux and Windows. In all of these environments, compilation happens successfully, and no warnings are generated during compilation.

I also added and enabled a 'COORDGEN_RIGUROUS_BUILD' macro which enables "/WX" on Windows and "-Wall -Wextra -Werror" on Linux and Mac. On the windows side, the level of warnings reporting is not increased (/W4 or /Wall) because in those cases the compiler starts reporting warnings in boost.

The PR contains a lot of changes, most of them related to type casting, but also some which correct minor bugs (i.e. using a float variable to get an element of a vector, assigning int values to floats and vice versa).