Update Travis CI builds (#33) for change in default Travis Ubuntu machine
Travis CI recently changed its default ubuntu machine from Trusty Tahr (14.04) to Xenial Xerus (16.04), probably following end-of-life for Trusty.

This meant changes to the default boost and compiler versions. While gcc/g++ 4.8 is still available, the new default boost version (1.58) seems to fail linking to it.

This PR updates the .travis.yml file to continue building on the three last Ubuntu LTS releases: Trusty (it still can be used, but it has to be explicitly specified, since it is no longer the default), Xenial (the new default), and Bionic Beaver (18.04, the last released LTS).

On each platform, the build is made using its default compiler and boost version (although the automatic apt installer for Bionic seems to be broken, so that we need a workaround).

Once Trusty is finally dropped by Travis, we can just remove that section. We also might add the next LTS once it is launched, probably next April.


This includes a work-around for a workaround for a travis issue with apt in bionic. Once the "sudo: unable to resolve host" lines go away from the bionic travis build, we should use idiomatic travis commands on bionic, as well.