avoid setting bonds info when reading in atom stereochemistry (#32)
writeStereoChemistry() was setting bond direction flags (inaccurately cause at this stage atoms usually have no coordinates) that are not cleaned up and can end up messing the stereochemistry of the final molecule. (bond directions are set at a later stage)

While this does not include a test for the issue, I've opened a case in our internal bug tracker to remind us to add a test.