Improve maeparser dependency handling (#41)
This aims at improving the way that the maeparser dependency is handled when building coordgen, and also makes it easier to include both in other projects. Also, it provides some documentation on how to build.
I brought in the CMake module we wrote to allow using find_package() to find maeparser compiled libraries, and also added another module to download maeparser from github if the library cannot be found.

Some details:

The maeparser_DIR option now points to the install path, instead of the .cmake files, which was a bit obscure if you don't know how CMake works (this effectively changes the recommended solution from #37).

If maeparser_DIR was specified (if it is not an empty string), maeparser's sources will NOT be used, and cmake will fail (since we expected to find a compiled library).

If there is a copy of maeparser at the download location (inside a maeparser-{version} directory in coordgen's source dir), the download will be skipped, and the already present sources will be built.

A specific branch/tag/commit may be specified to be downloaded. By default, the master branch will be used.

Travis and Appveyor scripts have been updated, since we no longer need to build maeparser in advance.