Don't copy vector when not needed.
The copy of the bonds vector was, seriously, 10% of total
duration for some example molecules that I looked at.
(because of copies that occured while checking whether
some atom is terminal)

Also fixes an extra copy for the ring vector, but that
doesn't have an impact that I know of. It's just odd.