Issue #52: Add a test that templates are used
Adds a test that when coordgen is run on the template files,
the templates are actually used. Does this by reading the
template files into molecules and then running
generateCoordinates(). It looks like the template matching code
adds the "rigid" property to any atom that had matched a
template, so that's what is checked here.