Bundles templats and allow coordgen to run without maeparser.
Now templates are always bundled into the coordgen library,
so there is no search on the file system. coordgen can also
be built without maeparser support (USE_MAEPARSER=OFF). This
disables user-local template files, but it allows coordgen
to be built without reliance on the maeparser library
(and therefore without boost).

This is useful for a couple of reasons:
* We definitely see bugs in template discovery
* 3rd party packagers seem to have trouble with
  building maeparser.
* Allow use in other novel places that can't use
  the boost:: infrastructure required by
  maeparser.

maeparser is always required for running the unit tests.

Includes a short Python script for generating the C++ files
describing the templates. This Python script currently requires
use of the Schrodinger Python tools, so it will need to be
manually updated if the templates are updated. There is a test
that will fail if a new template is added but isn't added to
the compiled templates.

run it like:

    $SCHRODINGER/run mol_generator.py templates.mae