CRDGEN-255: Penalize flipping constrained fragments (#93)
Formerly, if a caller used a template molecule to set the positions
of the atoms via constraints, there was no guarantee that chunks
within the template wouldn't be "flipped".

This is because coordgen uses molecule fragments to generate
geometries, and before this commit, it allowed those fragments
to "flip" to avoid interatom clashes. This commit heavily
penalizes flipping constrained fragments - but it doesn't
completely forbid it, because sometimes it is necessary.

We may still see issues after this if the template covers an
incomplete fragment.